split test_graph.py into three files
